+++ src/app/api/models/task-status.ts
@@ -16,14 +16,14 @@ export type TaskStatusEnum =
   | 'assess_in_portfolio'
   | 'attention_required';
 
-export type TaskStatusUiData = {
+export interface TaskStatusUiData {
   status: TaskStatusEnum;
   icon: string;
   materialIcon: string;
   label: string;
   class: string;
   help: {detail: string; reason: string; action: string};
-};
+}
 
 export class TaskStatus {
   public static readonly STATUS_KEYS: TaskStatusEnum[] = [
@@ -238,25 +238,25 @@ export class TaskStatus {
   ]);
 
   public static readonly STATUS_ICONS = new Map<TaskStatusEnum, string>([
-    ['ready_for_feedback', 'fa fa-thumbs-o-up'],
-    ['not_started', 'fa fa-pause'],
-    ['working_on_it', 'fa fa-bolt'],
-    ['need_help', 'fa fa-question-circle'],
-    ['redo', 'fa fa-refresh'],
-    ['feedback_exceeded', 'fa fa-low-vision'],
-    ['fix_and_resubmit', 'fa fa-wrench'],
-    ['discuss', 'fa fa-commenting'],
-    ['demonstrate', 'fa fa-commenting'],
-    ['complete', 'fa fa-check'],
-    ['fail', 'fa fa-times'],
-    ['time_exceeded', 'fa fa-clock-o'],
-    ['assess_in_portfolio', 'fa fa-folder-open'],
-    ['attention_required', 'fa fa-commenting'],
+    ['ready_for_feedback', 'thumb_up'],
+    ['not_started', 'pause'],
+    ['working_on_it', 'bolt'],
+    ['need_help', 'help'],
+    ['redo', 'undo'],
+    ['feedback_exceeded', 'visibility_off'],
+    ['fix_and_resubmit', 'construction'],
+    ['discuss', 'question_answer'],
+    ['demonstrate', 'record_voice_over'],
+    ['complete', 'done_all'],
+    ['fail', 'close'],
+    ['time_exceeded', 'schedule'],
+    ['assess_in_portfolio', 'folder_open'],
+    ['attention_required', 'sms_failed'],
   ]);
 
   // Material icons used by newer UI elements.
   public static readonly STATUS_MATERIAL_ICONS = new Map<TaskStatusEnum, string>([
-    ['ready_for_feedback', 'thumb_up'],
+    ['ready_for_feedback', 'thumb_up_off_alt'],
     ['not_started', 'pause'],
     ['working_on_it', 'bolt'],
     ['need_help', 'help'],
@@ -268,7 +268,7 @@ export class TaskStatus {
     ['complete', 'done'],
     ['fail', 'close'],
     ['time_exceeded', 'schedule'],
-    ['assess_in_portfolio', 'folder_open'],
+    ['assess_in_portfolio', 'rate_review'],
     ['attention_required', 'sms_failed'],
   ]);